requirements.txt
openai>=1.50.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
pypdfium2>=4.30.0
Pillow>=10.0.0
pypdf>=4.0.0
